Lockless tlog:
This commit remove all locking mechanisms
using semaphore and shared_mutex.

Packets are distributed according to it's volume ID.
